+++ src/_doc/mod.rs
@@ -2,7 +2,7 @@
 //
 //! Extra documentation about the library.
 #![doc = crate::_doc!(modules: crate; _doc: examples, features, macros, nightly, vendored)]
-#![doc = crate::_doc!(br+hr)] // gives way to the first root module
+//#![doc = crate::_doc!(br+hr)] // gives way to the first root module
 //
 // #![cfg(any(doc, test))] // RETHINK
 #![cfg_attr(nightly_doc, doc(cfg(any(doc, test))))]